Safari-609.1.11.1/LayoutTests/js/dom/modules/module-execution-order-inline.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test module execution order between inlined module tags.');
window.count = 0;
// Module will be executed asynchronously.
window.jsTestIsAsync = true;
debug('Module is not executed yet.');
shouldBe("count++", "0");
</script>
<script src="../../../resources/js-test-post.js"></script>
<script type="module">
shouldBe("count++", "1");
</script>
<script type="module">
shouldBe("count++", "2");
</script>
<script type="module">
shouldBe("count++", "3");
finishJSTest();
</script>
</body>
</html>
